+++ Backend - IGDS - Java 17, Spring, Postgres/demo/src/main/java/com/example/demo/enumeradores/TipoUsuario.java
@@ -1,4 +1,4 @@
-package services.enumeradores;
+package com.example.demo.enumeradores;
 
 public enum TipoUsuario {
     DOADOR, BENEFICIADO, HOSPITAL;
