+++ Backend - IGDS - Java 17, Spring, Postgres/demo/src/main/java/com/example/demo/model/UsuarioEntity.java
@@ -1,26 +1,24 @@
-package model;
+package com.example.demo.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import service.enumeradores.TipoUsuario;
+import lombok.*;
+import com.example.demo.enumeradores.TipoUsuario;
 
 @Entity
 @Data
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(schema = "public", name = "usuarios")
+@Table(schema = "public", name = "cartao")
 public class UsuarioEntity {
+
     @Id
     @Column(name = "id_usuario")
     @JsonProperty("_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id_usuario;
-    @Column(name = "nomeCompleto", length = 255)
+    @Column(name = "nome_completo", length = 255)
     private String nomeCompleto;
     @Column(name = "login", length = 255)
     private String login;
@@ -33,4 +31,6 @@ public class UsuarioEntity {
     @Enumerated(EnumType.STRING)
     @Column(name = "tipo_usuario", length = 255)
     private Enum<TipoUsuario> tipo_usuario;
+
+
 }
