+++ Backend - IGDS - Java 17, Spring, Postgres/demo/src/main/java/com/example/demo/repository/UsuarioRepository.java
@@ -1,10 +1,11 @@
-package repository;
+package com.example.demo.repository;
 
-
-import model.UsuarioEntity;
+import com.example.demo.model.UsuarioEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Long> {
+
+
 }
