+++ Backend - Sangue Amigo/demo/src/main/java/com/example/demo/controller/UsuarioController.java
@@ -1,21 +1,13 @@
-package controllers;
+package com.example.demo.controller;
 
-import model.UsuarioEntity;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import repository.UsuarioRepository;
-
-import java.util.List;
+import com.example.demo.repository.UsuarioRepository;
 
 @RestController
 @RequestMapping
 public class UsuarioController {
-
     @Autowired
     private UsuarioRepository usuarioRepository;
-
-    @GetMapping("api/usuario")
-    public List<UsuarioEntity> buscarUsuario(){ return usuarioRepository.findAll();}
 }
