+++ Backend - Sangue Amigo/demo/src/main/java/com/example/demo/repository/UsuarioRepository.java
@@ -1,11 +1,10 @@
-package repository;
+package com.example.demo.repository;
 
-import model.UsuarioEntity;
+
+import com.example.demo.model.UsuarioEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Long> {
-
-
 }
