+++ Backend - Sangue Amigo/demo/src/main/java/com/example/demo/service/UsuarioService.java
@@ -1,0 +1,4 @@
+package com.example.demo.service;
+
+public class UsuarioService {
+}
