+++ Backend - Sangue Amigo/demo/src/main/java/com/example/demo/service/enumeradores/TipoUsuario.java
@@ -1,4 +1,4 @@
-package service.enumeradores;
+package com.example.demo.service.enumeradores;
 
 public enum TipoUsuario {
     DOADOR, BENEFICIADO, INSTITUICAO
