--- Backend - Sangue Amigo/demo/src/main/java/controller/UsuarioController.java
@@ -1,13 +1,0 @@
-package controller;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import repository.UsuarioRepository;
-
-@RestController
-@RequestMapping
-public class UsuarioController {
-    @Autowired
-    private UsuarioRepository usuarioRepository;
-}
