--- Backend - Sangue Amigo/demo/src/main/java/service/UsuarioService.java
@@ -1,4 +1,0 @@
-package service;
-
-public class UsuarioService {
-}
